Refactor RemoteCache key base class into wildfly-clustering-ee-hotrod.
