+++ clustering/web/hotrod/src/main/java/org/wildfly/clustering/web/hotrod/session/fine/SessionAttributeNamesKey.java
@@ -21,13 +21,13 @@
  */
 package org.wildfly.clustering.web.hotrod.session.fine;
 
-import org.wildfly.clustering.infinispan.client.Key;
+import org.wildfly.clustering.ee.hotrod.RemoteCacheKey;
 
 /**
  * Cache key for session attribute names.
  * @author Paul Ferraro
  */
-public class SessionAttributeNamesKey extends Key<String> {
+public class SessionAttributeNamesKey extends RemoteCacheKey<String> {
 
     public SessionAttributeNamesKey(String id) {
         super(id);
