+++ clustering/web/hotrod/src/main/java/org/wildfly/clustering/web/hotrod/sso/HotRodSSOManagerFactory.java
@@ -25,11 +25,9 @@ package org.wildfly.clustering.web.hotrod.sso;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicReference;
 
-import org.infinispan.client.hotrod.RemoteCache;
 import org.wildfly.clustering.ee.Batcher;
 import org.wildfly.clustering.ee.cache.tx.TransactionBatch;
 import org.wildfly.clustering.ee.hotrod.tx.HotRodBatcher;
-import org.wildfly.clustering.infinispan.client.Key;
 import org.wildfly.clustering.marshalling.spi.MarshalledValueMarshaller;
 import org.wildfly.clustering.web.cache.sso.CompositeSSOManager;
 import org.wildfly.clustering.web.cache.sso.SSOFactory;
@@ -52,10 +50,9 @@ public class HotRodSSOManagerFactory<A, D, S> implements SSOManagerFactory<A, D,
 
     @Override
     public <C, L> SSOManager<A, D, S, L, TransactionBatch> createSSOManager(SSOManagerConfiguration<C, L> config) {
-        RemoteCache<Key<String>, ?> cache = this.configuration.getRemoteCache();
         SessionsFactory<Map<D, S>, D, S> sessionsFactory = new CoarseSessionsFactory<>(this.configuration.getRemoteCache());
         SSOFactory<Map.Entry<A, AtomicReference<L>>, Map<D, S>, A, D, S, L> factory = new HotRodSSOFactory<>(this.configuration.getRemoteCache(), new MarshalledValueMarshaller<>(config.getMarshalledValueFactory()), config.getLocalContextFactory(), sessionsFactory);
-        Batcher<TransactionBatch> batcher = new HotRodBatcher(cache);
+        Batcher<TransactionBatch> batcher = new HotRodBatcher(this.configuration.getRemoteCache());
         return new CompositeSSOManager<>(factory, config.getIdentifierFactory(), batcher);
     }
 }
